Add files via upload
This is the Pixelcoin ERC20 token launched on solidity (coding platform) for ethereum. In order to buy the token, you need to type in the contract address: 0x5B38Da6a701c568545dCfcB03FcB875f56beddC4 in MetaMask, follwed by the symbol (PXL) and the number 18. The other releases are Pixelcoin Merge-Link Blockchain release 1, and a simulation blockchain for testing called Pixelcoin Blockchain. The goal of this project is to create new innovation in the blockchain space such as enhanced security and transaction times through the new Merge-Link protocol that's the core foundation of this ERC20 token.